mod: User Service 인터페이스 추상화
