+++ src/main/java/com/hyoguoo/paymentplatform/user/presentation/port/UserService.java
@@ -1,0 +1,8 @@
+package com.hyoguoo.paymentplatform.user.presentation.port;
+
+import com.hyoguoo.paymentplatform.user.domain.User;
+
+public interface UserService {
+
+    User getById(Long id);
+}
